requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
sqlmodel==0.0.14
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

celery==5.3.6
redis==5.0.1

python-telegram-bot==20.8
google-api-python-client==2.121.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

openai==1.12.0
google-generativeai==0.5.4

pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1

httpx==0.26.0
aiofiles==23.2.1
